+++ app/services/express/middlewares/passport/index.js
@@ -1,39 +1,6 @@
 import passport from 'passport';
 import auth from './auth';
-
-function createSocialNetworkRoutes(app) {
-    const socialNetworks = ['facebook'];
-    socialNetworks.forEach(function (network) {
-        app.get(`/auth/${network}`, setSocialAuth(network));
-        app.get(`/auth/${network}/callback`, setSocialAuth(network));
-    });
-}
-
-function handleLogin(req, res, next) {
-    passport.authenticate('local', function (err, user, info) {
-        if (err) {
-            return next(err);
-        }
-        if (!user) {
-            return res.redirect('/login');
-        }
-        req.logIn(user, function (err) {
-            if (err) {
-                return next(err);
-            }
-            return res.redirect('/');
-        });
-    })(req, res, next);
-}
-
-function handleLogout(req, res, next) {
-    req.logout();
-    next();
-}
-
-function setSocialAuth(provider) {
-    return passport.authenticate(provider, {scope: ['email']});
-}
+import {createSocialNetworkRoutes, handleLogin, handleLogout} from './utils';
 
 export default (app) => {
     app.use(passport.initialize());
